+++ Configuration/JavaScriptModules.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'dependencies' => ['backend'],
+    'imports' => [
+        '@oliverklee/seminars/' => 'EXT:seminars/Resources/Public/JavaScript/BackEnd/',
+    ],
+];