+++ Tests/Unit/Controller/BackEnd/RegistrationControllerTest.php
@@ -559,7 +559,7 @@ final class RegistrationControllerTest extends UnitTestCase
 
         $this->pageRendererMock
             ->expects(self::once())->method('loadRequireJsModule')
-            ->with('TYPO3/CMS/Seminars/BackEnd/DeleteConfirmation');
+            ->with('TYPO3/CMS/Seminars/BackEnd/DeleteConfirmationAmdModule');
 
         $this->subject->showForEventAction($eventUid);
     }